Rename filename for shorthand helper functions
