+++ packages/meteor-type-validation/src/server/ShorthandFunctions.ts
